requirements.txt
# Core LangChain dependencies

langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
langchain-core==0.3.80
langchain-huggingface==0.3.1
langchain-chroma==0.2.6

# Vector store and embeddings
chromadb==1.3.5
sentence-transformers==3.3.1
einops=0.8.1

# Document processing
cryptography
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
unstructured==0.11.8

# API framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# HTTP client for tools
httpx==0.26.0
requests==2.31.0

# Environment variables
python-dotenv==1.0.1

# Data processing
pandas==2.1.4
numpy==1.26.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

